tast-tests: add platform/memory_pressure.go
This is a rewrite of the platform_MemoryPressure autotest.
There are slight changes in functionality.
Most importantly, it uses Web Page Replay to serve all requests.

This test reproduces a streamlined user behavior, with the intent
of filling up RAM quickly in a somewhat realistic way.  It opens
a bunch of tabs from a list, going over the list multiple times
as needed.  After opening each tab, it cycles through a small
number of recently-used tabs.  While visiting each tab, it scrolls
the tab down and up for a short time.  After the first tab discard
(or OOM kill) it collects some stats, then it rests for a while
and collects some more stats.

BUG=chromium:921196
TEST=ran on x86 and ARM platforms

Change-Id: I6e778b4173ad4b43eeb2abff093cc462482f7c6d