tast-tests: Add security.StatefulFiles local test.
Port the security_StatefulPermissions Autotest test to
security.StatefulFiles. This test checks ownership and
permissions of files in /mnt/stateful_partition.

The old test takes the approach of using su to run many long
"find ... -writable" commands as different users. Files that
are expected to be writable are excluded using -prune, and
the remaining files are reported as errors. This typically
takes 45-60 seconds.

The new test instead does a single pass over the directory.
Patterns are supplied to explicitly match paths and check
their ownership and permissions, and fallbacks are used to
check that unmatched files are only writable as root. This
takes 1-2 seconds.

The old test appeared to list some outdated paths, which
I've removed. I've also noticed differences across the
various lab DUTs that I've run this on, which I've tried to
permit. Nevertheless, I suspect that there will be many
errors that I'll need to address in a followup change.

BUG=chromium:877733
TEST=updated unit tests and ran new test against DUTs

Change-Id: I3ca14c13cb4dfc20703be230ef0afc8431ee8130
Reviewed-on: https://chromium-review.googlesource.com/1335167
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>