Update patch set 12
Patch Set 12:

(2 comments)

Thanks for the review.

nya@,
  Here's a comment for you in previous patch. I'm not sure if you see this.

I found there's a caveat in the auto-initialization line
https://cs.corp.google.com/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/perf/perf.go?l=84

Since perf.Values is encouraged to be passed by value, if it's passed multiple times, the initialization only initializes the last copy. For example, if I write "var perfValues perf.Values" in diskIOPerf(), then pass it to runJobFile(), then pass it again to reportMetric(). Only the copy in reportMetric() got initialized. So the best we can do is still initialize it in diskIOPef() manually.

The auto allocation did nothing wrong, but I think it kind of covered the un-initialization problem. Users should be encouraged to allocate memory for a map by themselves. In case like this there wouldn't be a runtime "assignment to a nil map" error so user might ignore the problem at the first place.
An alternative is to make perf.Values a struct and provide a New() so user always get an initialized object. Thus one wouldn't need to know the internal of perf.Values.

Patch-set: 12
Reviewer: Gerrit User 1002188 <1002188@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>