Update patch set 1
Patch Set 1:

I like short names.  But, that doc says:

"A rule of thumb
The greater the distance between a name's declaration and its uses, 
the longer the name should be."

I think a better phrasing would be

"The greater the distance between a name's declaration and the end of its scope, ..."

because when the scope is not short (as is the case here) then it's harder to tell if there are uses beyond the ones close to the declaration.

Patch-set: 1