tast-tests: Use auto disk type in Crostini tests
Switch the disk image type to DISK_IMAGE_AUTO, which will let concierge
select the best format for the kernel and filesystem (raw on systems
that support FALLOC_FL_PUNCH_HOLE, qcow2 on older kernels that don't
support it).

The Chrome browser code and concierge_client/vmc tools have already
switched to using auto, so we should make the tests use it as well to
test what we are shipping.

BUG=chromium:893380
TEST=tast run vm.CrostiniStartEverything on nami and kevin

Change-Id: I79ac332225ecd6b379fe2859afe9bdcd01784dad
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>