video.WebRTC*: Don't load/unload vivid in Tast tests
Since vivid is loaded by upstart service on VM boards, each test
case doesn't have to load/unload it any more.

But, we still need to check if the device is VM or not. It's because
we check the ratio of broken frames when the used camera is vivid,
which doesn't have hardware flakiness.

BUG=chromium:904730, chromium:899595
TEST=tast run video.WebRTCCamera on amd64-generic VM
TEST=tast run video.Capability after running WebRTC tests on VM
CQ-DEPEND=CL:1333099, CL:1333102

Change-Id: I2e794a72dc1f826f887c8489a03c7a5206f1226c
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>