tast-tests: Fix syslog, cups in security.StatefulFiles.
Update security.StatefulFiles to permit the contents of
/mnt/stateful_partition/encrypted/var/log to be owned by the
syslog group. This group appears to be used for atrus.log
and hammerd.log, and it seems safe to permit (we already
permit logs to be owned by the syslog user, and we make sure
they aren't group-writable).

Also check that the cups user exists before adding a rule
for /mnt/stateful_partition/encrypted/var/spool/cups. This
user apparently doesn't exist on veyron_rialto (even though
it's a member of the lp group there).

Finally, skip
/mnt/stateful_partition/unencrypted/preserve/cros-update.
This directory looks like it's cros_update.py script, so I
think that it's testing-specific, and we already skip other
files in .../preserve/log.

BUG=chromium:877733
TEST=ran security.StatefulFiles successfully on soraka and
     veyron_rialto

Change-Id: If9c9c09d9e9b88c57d415c21f260511d74e3c821