Spoken feedback events test using for ARC.
An Android application containing a toggle button and checkbox is
started, and spoken feedback is used navigate and interact with these
elements. Apk will be uploaded to cloud storage at a later time,
similar  to how it is done in another tests ( see media_session.go)

time.after is required in between keyboard to ensure that
spoken feedback (ChromeVox) has a chance to read out elements,
before processing the next event.

A helper function, connecting to a specified Chrome extension was also
added in chrome.go.

TEST=tast run DUT_IP arcapp.AccessibilityEvent
BUG=b:118784915
Change-Id: I2d8b268910f1e3fd597746d3da59a8ef6de477c2