tast-tests: Take ownership of TPM before testing.
Update the "ready" package to attempt to ensure that the TPM
is owned before testing starts.

Also make it wait for the cryptohomed process to be up for
at least 10 seconds before running the cryptohome command to
try to avoid triggering hangs that we've seen just after
startup in the past.

Also avoid waiting for Upstart jobs that aren't installed.

BUG=chromium:901363
TEST=ran "crossystem clear_tpm_owner_request=1" and rebooted
     to clear ownership, then ran ui.ChromeLogin and
     verified that ownership was retaken before the test was
     run. also verified that ownership is skipped on a VM

Change-Id: Ib205dc268b5e4bea6ea13d77e86037440740c102