Update patch set 2
Patch Set 2:

LoggedIn tests are failing sometimes (flaky), but it looks like flaky of UI. So it should be addressed separately, and that part of this CL LGTM.

Guest tests are failing, too, because some files seem to be deleted during the run.
I think it will be fixed by crrev.com/c/1349630 (as CQ-DEPENDed), but the missing files I found are /home/chronos/Default/MANIFEST-*, and /home/chronos/Default/*.log, which are not in the commit message of the CL, so I'd like you to double check if it is actually OK to be skipped, just in case.

Patch-set: 2
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>