tast-tests: Fix test failure with installing packages
There was a new failure where installing from a .deb file would fail a
lot of the time. This was because we added a time sync test which was
setting the clock backwards to the epoch and then to the current time.
That would then end up firing the apt-daily task which would interfere
with the package install. We now just set the clock back one hour and
then sync it, which won't fire the daily task.

BUG=chromium:908527
TEST=tast run vm.CrostiniStartEverything

Change-Id: I942a2ee7524c789791a3c0e75eddc1e238e03bc4