tast-tests: Make chrome.New check that cryptohomed works.
Add a new cryptohome.CheckService function that chrome.New
calls early on to check that cryptohomed is running and that
its D-Bus service is registered. If it's not running,
additional information is collected about the state of other
daemons that can cause issues (attestationd, tpm_managerd,
trunksd, and chapsd).

This is intended to improve the error messages that are
reported in situations where login would fail later anyway
due to pre-existing issues on the DUT.

BUG=chromium:910407
TEST=manual: hacked the code to always report failure and
     then ran ui.ChromeLogin against betty (fake TPM),
     caroline (TPM 1.2), and eve (TPM 2.0) DUTs and verified
     that errors were reasonable

Change-Id: I415151d07303d87990c80ddd8bc9ca9fa3bb88fd