tast-tests: Add dev_image_old to security.StatefulFiles.
I'm not sure what creates it, but per the failure at
http://stainless/browse/chromeos-autotest-results/263445795-chromeos-test/,
it looks like a /mnt/stateful_partition/dev_image_old
directory can be left behind on DUTs. Skip it as is already
done for the dev_image directory.

Also add some more blank lines between groups of patterns to
make the code easier to read.

BUG=chromium:877733
TEST=manual: ran the test

Change-Id: I1eafbee88dce4517ad615d9785f3da0fb5e62045