Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > I am a little bit concerned about if a new code reads entire the yuv content in memory. The size can be up to 5MB, iirc. I would like to leave derat@ for the decision.
> 
> could something like https://golang.org/pkg/io/#MultiWriter be used to both send the data to hash.Hash and to pass it to createYUVFile?

What I was trying to do is to avoid writing a file until it is actually usable (i.e., after md5 sum check and file format conversion), in order to simplify the I/O error handling.

Is 5MB in a temp buffer in a function really concerning size?
If so, it is necessary to combine "writing a file" and "preparing the data" which are split by this CL.
Because it'd introduce complexity, I'd rather not to do that, unless 5MB is concerning.

WDYT?

Patch-set: 2
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>