security.StatefulPartitionHardening: relax test conditions
While porting this test I strengthened one of the conditions, but
it seems like this broke the test on some platforms. Bring it in
line with its autotest counterpart.

(Notably, FIFO access in /home seems to be inconsistent across
boards. We should make this consistent and then put /home into the
approriate category.)

BUG=None
TEST=ran successfully on nautilus

Change-Id: I8351af0fafc71e359edfc021c3b1fd87817aff0f