Update patch set 3
Patch Set 3:

We want to have the long list here if possible...

IIUC the problem we want to catch with the test is to forget setting first_api_level explicitly for new boards. Is it possible to set the "default" first_api_level to something weird (like "1") and detect it in the test?

Patch-set: 3
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>