Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> LGTM.
> 
> (off topic): Do you know how GPU sandbox is disabled on VMs?
> We are disabling the gpu_sandboxed tests on VMs, but if there is something (like a flag etc.) to control it, it's nice to be looked at for the purpose, I thought.

i'm not sure. from the exception in the bug, it looked like the issue may be caused by GPU information not being available when running in a VM, but chrome://gpu also said that sandboxing was disabled when i looked at it on a betty VM.

kenneth, do you happen to know what the story is around GPU sandboxing when running chrome os in a VM (or who would know? looks like a lot of people are out for the holidays)?

is it worthwhile trying to make the isSandboxedForTesting() helper function report true in this case (assuming that sandboxing actually works in a VM)?

Patch-set: 1
CC: Gerrit User 1105344 <1105344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>