Update patch set 5
Patch Set 5:

(2 comments)

> Patch Set 3:
> 
> We want to have the long list here if possible...
> 
> IIUC the problem we want to catch with the test is to forget setting first_api_level explicitly for new boards. Is it possible to set the "default" first_api_level to something weird (like "1") and detect it in the test?

Unfortunately, many boards doesn't have special config for overwriting this value yet... so many of them depend on inheriting the default value (nocturne is an example - when we need to overwrite the value from the default 25, there is no special config that we can use to rewrite it yet... so we are forced to hack the build sh script by adding "if nocturne, set first_api_level to 28").

Having board config's overwrite for each board is a much bigger task which might or might not have enough benefit ~ so I believe having this test is unfortunately more practical (at least for the time being).

Patch-set: 5
Reviewer: Gerrit User 1221932 <1221932@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>