tast-tests: Be sure that we're mlocking memory in Chrome
For consistent performance under memory pressure, we try to mlock pieces
of Chrome. This tast ensures that this mlocking still happens in the
future, since the only indication we otherwise have of it disappearing
is a degraded UX and a few LOG(INFO) messages.

BUG=chromium:913700
TEST=tast run $(cat /tmp/cros_ip) platform.ChromeMlocked

Change-Id: I5bc90effc49e562cbf7e1770238cc10afa2ba5c6
Reviewed-on: https://chromium-review.googlesource.com/1381891
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>