video: Add functions to Mute/Unmute not to make a sound during testing
Videos in some tests has an audio. It doesn't matter whether DUT sounds or not.
It is convenient for developers that test automatically makes DUT mute during
the test.
Note that it matters that chrome processes for sound, because not only we want
to measure metrics in chrome standard video playback but also chrome video
playback behavior is changed by the existence of sound in the video.

BUG=chromium:890733, chromium:911095
TEST=tast run video.Play*

Change-Id: Ibe4895c1cfc6991beaf5672724bb8c9bf15f194c