video: Add functions to Mute/Unmute not to make a sound during testing
Videos in some tests has an audio. It doesn't matter whether DUT sounds or not.
It is convenient for developers that test automatically makes DUT mute during
the test.
Note that it matters that chrome processes for sound, because not only we want
to measure metrics in chrome standard video playback but also chrome video
playback behavior is changed by the existence of sound in the video.

BUG=chromium:890733, chromium:911095
TEST=tast run video.Play*

Change-Id: Ibe4895c1cfc6991beaf5672724bb8c9bf15f194c
Reviewed-on: https://chromium-review.googlesource.com/1390677
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>