security.ASLR: Relax test condition to match autotest
The test condition was accidentally strengthened during porting
so that the test would fail if any randomized address was the same
as the first iteration, rather than if all of them were. This
caused the test to flake unnecessarily.

This also does some other cleanup like combining two functions that
were only ever used in composition.

BUG=chromium:915824
TEST=run test on nautilus

Change-Id: I578dc454e38ff60fea8a6635f64a2c43488567ae