arc: Stop using --arc-start-mode.
The flag is being deprecated, so stop using it.

We started using the flag to avoid the Android mini container
killed during session login (crbug.com/872135), but apparently
the problem went away somehow. Let's just remove it and see if
this is okay by watching how the arc.MiniContainer test goes.

BUG=chromium:919445
TEST=tast run DUT arc.MiniContainer  # w/ nocturne and betty

Change-Id: I38ca1fd9f888e65334fab3c620cb007fb032caf3