Update patch set 13
Patch Set 13: Code-Review+2

> Patch Set 13:
> 
> PTAL. Sorry, it is from last month. CQ gives error for missing "android". So "android" is added in SoftwareDeps.

hmm, i can't see the failure, but i assume that this is tast/local/bundles/cros/arc/registration_test.go, right?

this seems like it's a bug in the unit test. iirc, "android_p" implies "android", so tests should be able to specify either. nya@ might have ideas about the best way to fix this, but probably testcheck.SoftwareDeps needs to be updated to support "any of" in addition to "all of".

adding "android" seems okay as a workaround if you don't want to wait for the unit test to be fixed, but if so, please add a TODO to remove it later. thanks!

Patch-set: 13
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2