Update patch set 2
Patch Set 2:

(2 comments)

Thanks. Honestly I didn't realize a review would be triggered but thank you for doing so!
I'd be fine with reverting this and adding it to CheckProcesses but I want to be able to verify both the parent and child are up. Do you prefer I modify CheckProcesses with a count or continue with this change? Thanks

Patch-set: 2
Reviewer: Gerrit User 1307765 <1307765@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>