tast: Don't wait for keyboard node when checking visiblity.
Sometimes tast.ui.VirtualKeyboardOmnibox times out when
checking if the virtual keyboard is initially visible or not.
In the current code we wait for the keyboard node to be
present in the automation tree before checking visibility.

However, this is causing timeouts sometimes, so we'll just
change it so that if the keyboard is not in the automation
tree then the keyboard is considered to be visible.

TEST=eve
BUG=chromium:893957

Change-Id: I7f3b34b5b1e7a5f31245fe4e6ca30b95a7cde1df