tast-tests: Add support for creating virtual input devices.
Update the input.Keyboard function to automatically create a
virtual keyboard device using uinput if a physical keyboard
is not present. This should make it possible for tests that
require a keyboard to run on Chromeboxes or tablets that
don't have keyboards.

Also rework some of the types exposed by the input package's
internal ioctl code to be closer to the types used by the
golang.org/x/sys/unix package's ioctl code.

Finally, make the example.Keyboard test non-disabled and
make it sleep several seconds after opening the keyboard to
give Chrome time to recognize the device before events are
sent. (I haven't found a better way to automate this yet.)

BUG=chromium:915965
TEST=manual: forced virtual keyboard to be created using
     methods for both old and new kernels and verified that
     evtest and chrome receive events sent by test

Change-Id: Ib31cd10fbc0c723b54ed3668bd9fb973d298ca1b