Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 3:
> > > 
> > > (4 comments)
> > > 
> > > Thanks!
> > > 
> > > And I have some irrelevant questions: 
> > 
> > the list at go/tast-users may be better for random questions since others are more likely to answer too. :-)
> > 
> > > 1. How can we create the same test with different parameters?
> > > 
> > > For example, we want to use ALSAConformance.InternalSpeaker, ALSAConformance.Mic for testing different nodes. In autotest, we can set config.InternalSpeaker and config.Mic to do it.
> > 
> > please put the common code in a subpackage, i.e. a "conformance" or similar package under the "audio" directory. it can export a Run or similar function that each of the conformance tests calls with its own test-specific configuration. then you'd have a separate short audio.ALSAConformanceSpeaker, audio.ALSAConformanceMic, etc. tests for each configuration.
> > 
> > see the video tests for extensive use of this pattern.
> > 
> > > 2. How do we set the dependency with labels in test labs?
> > > 
> > > I want ALSAConformance.Mic test only run on the DUT which has audiojack. We can check the label "test_audiojack" from labs to do it (Like DEPENDENCIES field in autotest config file). Can we also set labels dependency in tast?
> > 
> > see go/tast-deps and the SoftwareDeps field. in particular, the existing audio_play and audio_record features may be useful here.
> 
> Thanks. But by tracing codes, SoftwareDeps seems just a fixed config for each device. It extracts flags from /usr/local/etc/tast_use_flags.txt, which has not been changed since it was created.
> 
> In test labs, we might have different configure for the same devices. For example, some have audiojack plugged in but the others not. These DUTs are distinguished by label "test_audiojack" in labs. Do we have any way to get and use these labels?

i see. yes, it's correct that SoftwareDeps is primarily used to check for software features that are a function of the DUT's system image. these features don't differ across different DUTs in the lab that are all using the same board.

autotest labels are different, in that they're statically assigned on a per-DUT basis and stored in the autotest database. they should be used for scheduling decisions (i.e. if you have a test depends on a chameleon device being attached, it needs to be scheduled on DUTs that have a chameleon and not ones that don't). tast doesn't have control over where its tests are run -- that's still autotest's job, since it manages the lab.

i don't know exactly what audiojack is, though. is there a doc with more info? does it use servo?

and is audiojack only attached to DUTs in a specific pool, or is it more-or-less arbitrarily distributed across DUTs in the main pool(s)?

> 
> > 
> > > Thanks for your assistance.

Patch-set: 4
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>