tast-tests: Add telemetry.WilcoTelemetry
Add integration tests for the "telem" command line tool that pmoy@ has
been writing. This series of tests calls telem with all of the currently
implemented items and groups. telem then communicates with libtelem, diagnosticsd,
and procfs, and parses the results. Each of these systems has its own
unit tests, so this is just an end-to-end test to ensure all the
plumbing is is working between them

This is currently only relevant on the Wilco platform, but is planned to
be generalized later for telemetry work on future platforms as well.

BUG=chromium:901126
CQ-DEPEND=CL:1377689, CL:1377550
TEST=ensure that telem is installed on device:
> emerge-sarien diagnostics
> cros deploy {DUT_IP} diagnostics
and then run test on host
> tast -verbose run {DUT_IP} telemetry.WilcoTelemetry

Change-Id: I5489e8136e8941177fae24d1d084ff812fff34a1