Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> update:
> 
> if I add a rule to skip the "cros_ec" device, then it will create the virtual keyboard with the correct attributes. So, regarding my 64 vs 32 bit concern, it works Ok.
> 
> But with the virtual keyboard, the Keyboard test hangs. Same problem as with Nocturne.

thanks for all the details! i'll look into skipping cros_ec, but huh, that's weird -- the appearance of it seems very similar to a real keyboard. :-/

regarding the timeout, does it help if you increase the length of the sleep between creating the keyboard and typing? it's surprising to me if 5 seconds isn't enough, as that works on the very slow VM that i'm using. if chrome doesn't log any info about new keyboard devices, i might need to dig into its input-device-tracking code. i probably won't be able to do that until the end of the month, though.

and one more question, if you're still set up to test: if you run "evtest" in a terminal on the DUT after the device is created but before the key events are injected, can you see the events there? just want to make sure that the problem isn't happening at a lower level for some reason...

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>