Add tast test to check accessibility tree in ARC app.
Ensure that accessibility tree is as expected.
Obtain the entire accessibility tree, then check that tree exists, and
check that contents is as expected.
Some lines in the tree have values at the end of the line which change
with each run, so the string comparisson uses .contains.

The expected tree is in the file "data/accessibility_event_expected_tree.txt".

TEST=tast run DUT_IP arcapp.AccessibilityEvent
BUG=b:118784915

Change-Id: I2b6b20c2d8908401b1a459834953e9fa7ae873bc