Update patch set 3
Patch Set 3:

uploaded a patchset that fixes the main issue.
now the virtual keyboard is being recognized by Chrome.

the problem was that, apprently, the bus type was incorrect. it seems that the fields on devID were in the incorrect order.

Patch-set: 3
Reviewer: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>