tast: Allocate time for rolling back to clamshell
When a funtion call in the middle of arc.ChromeVirtualKeyboard was timed
out, inject_powerd_input_event didn't run, because the context was
already timed out. This affected another test, arc.PhysicalKeyboard.

In order to avoid that, we should use shortened timeout for the body of
the test.

TEST=tast eve-DUT arc.ChromeVirtualKeyboard
BUG=b:122457709
Change-Id: I6356440390cdeacdb79a38007eea35305f68866b
Reviewed-on: https://chromium-review.googlesource.com/1429304
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>