tast: Add custom timeout to WaitForExists.
In arc.ChromeVirtualKeyboard and arc.PhysicalKeyboard, GetText() was
called when WaitForExists() timeouts, in order to give a better error
message. However, GetText() always failed in a such case, because
context was already timed out. To avoid that, we have to give
custom shorter timeout for WaitForExists().

Creating a child context with shorter timeout didn't work, because
UiAutomator still got disconnected when doing that.

TEST=tast run eve-DUT arc.ChromeVirtualKeyboard
TEST=tast run eve-DUT arc.PhysicalKeyboard
BUG=b:122456478
BUG=b:122457709

Change-Id: I953402fa1c89eecccde3eb3779cf3291c77a4d02