Update patch set 7
Patch Set 7:

(4 comments)

Thanks for the great reviews.  All done I think except I am not sure if making Close() idempotent is stylistically acceptable.  The problem is that I don't know if the kernelmeter in the test is open or already closed when a fatal error occurs.  I could export a method to query if the kernelmeter is running, but this seems simpler if acceptable.

Patch-set: 7
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>