tast: Add VirtualKeyboardSuggestions test.
The virtual keyboard gets its word suggestions from a service
called the decoder. However, this decoder is built elsewhere, so
we don't really have any tests that make sure the decoder works
with the UI.

This test ensures that the virtual keyboard does indeed load the
decoder correctly and that the decoder can provide suggestions to
the virtual keyboard.

There is a quirk with decoders however. Decoders ignore what the
user types until it is fully loaded. There is no easy way to wait
for the decoder to load, so we just have to keep pressing keys
until the decoder provides suggestions. Until we have a good way
of waiting for the decoder to load, we'll just do this hack.

BUG=chromium:879073
TEST=veyron_minnie

Change-Id: I2893e3e108fd76371cf31f7f996c863038a0cecb