tast-tests: Make ui.MashLogin depend on "stable_egl"
The test crashes in the nvidia driver on nyan variants. We don't
know if this is due to chrome or due to the driver, but for now
just skip the test like we do in the autotest version of this
test.

BUG=chromium:717275
TEST=tast ui.MashLogin
CQ-DEPEND=I1b9b0190fbf95b958157d77b5b4eac8240fe3c32

Change-Id: I5e2b40115020dd85248ca2fb89c9e93a8f8e4e21