Add tast test to check accessibility tree in ARC app.
Ensure that accessibility tree is as expected.
Obtain the entire accessibility tree, then check that tree exists, and
check that contents is as expected.
String comparison uses .contains, to account for unique URL for ARC app in accessibility tree.

The expected tree is in the file "data/accessibility_tree_expected.txt".

TEST=tast run DUT_IP arcapp.AccessibilityTree
BUG=b:118784915

Change-Id: I7439e1306ec3f357b932b7a7cf80ee51673308dc