platform_MemoryPressure enhancements
This enhances the test in various ways.

First, to stimulate the loading and other page activity,
each page is "wiggled" i.e. scrolled back and forth once
when visited.

Second, we compare tab switch times in the first set of
working tabs under light memory pressure (at the beginning
of the test) and heavy pressure, at the end.  The difference
can be an indicator of user experience under pressure.

We go through some trouble to ensure that the initial set
of tabs is not discarded.  We periodically revisit them
as we open tabs, to force them to the top of the LRU list.
(Or the bottom, depending on how you see it.)  We also pin
them using chrome.tabs.update().

BUG=chromium:921196
TEST=ran the test many times, checked outputs
CQ-DEPEND=CL:1436885

Change-Id: Iff2852c317799947fd0a250e86f8f6e0d2c330fe
Reviewed-on: https://chromium-review.googlesource.com/1437485
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>