platform_MemoryPressure: measure tab switch times after discard
At the end of phase 1 (first discard) the system is believed
to be under pressure.  We measure tab switch times in phase 2
as an indicator of user experience under pressure.

To maximise reproducibility and the ability to make meaningful
comparisons, we perform the tab switches on the initial set of
opened tabs.  To ensure they are not discarded, we pin them
using chrome.tabs.update().

BUG=none
TEST=none

Change-Id: Iff2852c317799947fd0a250e86f8f6e0d2c330fe