Fix arc.MediaSessionGain test
We made a change to audio focus so that audio
focus requests lose a different type of focus
when they lose to another kGain type. This broke
the tast test as it used the wrong enum value.

Also needed a fix to not check the result text
box in the test app which is not updated when the
lost focus type is -1.

TEST=Run tast
BUG=b:123412387

Change-Id: I04423f2bce0a79c98325aff4c1a15ffbd477e212