tast tests: arc.MidiClient: Read & check MIDI message
This CL updates the test to extract the relevant "Midi Through" port
number using arecordmidi, and then record messages being written to it
using the same utility.

Once the messages are successfully recorded, they are inspected to make
sure the MIDI message that was written from the ARC++ app was
successfully received.

BUG=b:122860765
TEST=tast -verbose run -build=true pmalani-kevin arc.MidiClient passes

Change-Id: I3a38e88fe53cc6a2e4a28ed4dbf3623813582163