tast tests: arc.MidiClient: Read & check MIDI message
This CL updates the test to extract the relevant "Midi Through" port
number using arecordmidi, and then record messages being written to it
using the same utility.

Once the messages are successfully recorded, they are inspected to make
sure the MIDI message that was written from the ARC++ app was
successfully received.

Also, this CL renames the test to MIDIClient (which is more in
accordance with the Go style guide).

BUG=b:122860765
TEST=tast -verbose run -build=true pmalani-kevin arc.MIDIClient passes

Change-Id: I3a38e88fe53cc6a2e4a28ed4dbf3623813582163
Reviewed-on: https://chromium-review.googlesource.com/1437577
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Prashant Malani <pmalani@google.com>