Update patch set 3
Patch Set 3:

I believe the original code was correct.

By running "defer d.Close()", d.Close() is not called, but it is queued for the end of the function. In this case, it is called when "return err" or "return nil" is hit.

And what d.Close() does is to stop the UI automator server running in the Android container. It does not stop the Android app being manipulated.

So I believe the original code was correct. Could you elaborate on what you think is the cause of the flakiness?

Patch-set: 3
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>