Update patch set 3
Patch Set 3: Verified+1

Thank you.  The comments on options are all correct---it's a lot of boilerplate.

This is a much simpler solution, which obviously doesn't scale, but probably it doesn't need to, and if it does we can worry about it later.

If you think this is not generally useful, I don't mind implementing it in my test only.

Patch-set: 3
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1