Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> To be honest, I think naming individual contacts is a bad idea (it introduces single points-of-failure). If you desperately want to add an individual, adding me is fine, but folks will have a significantly easier time contacting the list than contacting me directly.

while i think a list is useful as a backup, i still feel like it's best to recommend an individual. when a group of people is responsible for something, it's sometimes the case that nobody is responsible for it. :-(

Patch-set: 3
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>