Update patch set 1
Patch Set 1:

are there already fixes in the pipeline for the current failures?

http://stainless/search?view=list&first_date=2019-01-24&last_date=2019-01-30&builder_name=-release%24&branch=%5Emaster%24&suite=%5Ebvt-perbuild%24&test=%5Etast%5C.platform%5C.CrosConfig%24&status=WARN&status=FAIL&status=ERROR&exclude_cts=false&exclude_not_run=true&exclude_non_release=true&exclude_au=true&exclude_acts=true&exclude_retried=true&exclude_non_production=true

also, this test is verifying different data on every board and/or model, if i understand correctly. the CQ doesn't have that kind of coverage, so the test is likely to be broken on some devices at any given time even if it's passing on all of the hardware that's used by the CQ.

Patch-set: 1
CC: Gerrit User 1206284 <1206284@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1268042 <1268042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>