tast-tests: Fix flakiness for AccessibilityEvent
Wait for ChromeVox to stop speaking before proceeding to the next event.

Although keyboard events were injected the action was not propagated to the element.
This meant that the correct view was not receiving focus, or that although the view had focus activation event (search+space)
was not being applied to the element.

In the first case, the focus gets stuck at the view. In the second case, although the correct
element has focus, the search+space action does not get applied to the element (togglebutton).

By ensuring that ChromeVox has stopped speaking, we can proceed to the next event
(interacting with ChromeVox, or pressing a key event).

Locally passses on:
- Grunt/ Careena (R74-11679.0.0)

Test=tast run DUT_IP arc.AccessibilityEvent
BUG=b:123397142

Change-Id: If6a01c1efaa5841ce56db8ce8117501e0d7e932e