perf: Introduce a constructor of Values.
It has been a source of confusion and bugs that Values is a map;
for example, we had a bug that "var p perf.Values" causes nil
deference (already fixed), and it's not consistent whether to use
perf.Values by value or pointer on passing to function arguments.

This CL tries to solve the problem by making it a struct, and
introducing NewValues constructor.

BUG=None
TEST=fast_build.sh -T

Change-Id: I5e1de669b7bbe0305d135e2040bfaac20566dfe0
Reviewed-on: https://chromium-review.googlesource.com/1457565
Commit-Ready: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>