Update patch set 1
Patch Set 1:

This is in preparation to running in the nightlies.  Without this, the test takes too long on 8 and 16 GB devices, and I don't think it does anything useful most of the time.

Also, the refresh of the initial tab set is a source of considerable pressure, so the test now does it at every new tab, instead of every 5.

Finally, tab discards can take a while, so the test now waits a bit after noticing the first discard before it collects stats.

Patch-set: 1
Reviewer: Gerrit User 1000028 <1000028@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>