printer: Fixing module installing and adding module cleanup
The modprobe function requires the "-a" option in order to install
multiple modules at the same time so this change fixes the issue in
InstallModules. This change also introduces the RemoveModules function
which is used to uninstall the usbip kernel modules after the test has
completed.

BUG=None
TEST=Able to run AddUSBPrinter test successfully using tast command

Change-Id: I150a245e4ecf376eb320ecbe975bd7036ee32fa6
Reviewed-on: https://chromium-review.googlesource.com/1465191
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>