tast-tests: Remove chrome_login dep from security.Mtab.
Remove a "chrome_login" dependency from the security.Mtab
test. This test doesn't actually exercise Chrome; it just
restarts the ui job to clean up stale mounts. We can
dynamically check for the existence of the ui job at runtime
instead.

This test is failing on the Chrome CQ since it remounts the
root partition read/write to update Chrome. Removing the
dependency should make it not be run there.

BUG=chromium:931431
TEST=manual: ran against caroline and panther DUTs

Change-Id: I16af921b133398e828eefa7598aec873adaca007