Update patch set 3
Patch Set 3:

note: I found the reason why we cannot use 'nil' for env in RunAsync.

If nil is used for env, the current enviroment variable (= os.Environ()) will be used.
(cf. https://golang.org/pkg/os/exec/#Cmd)
But, we cannot use os.Environ() for the binary tests because of the same reason with
https://chromium.git.corp.google.com/chromiumos/platform/tast-tests/+/09b4f053c758fae904cba538dce50ead0d165ee7/src/chromiumos/tast/local/chrome/bintest/bintest.go#73

Patch-set: 3